+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.0.2
+- Add new feature to redact password, key, secrets from HTTP query parameters
+- Upgrade iOS agent to 1.6.3 which fixed duplicated beacons issue
+
 ## 3.0.1
 - Remove print in code since it conflicts with some linter rules
 - Upgrade iOS agent to 1.6.2 which improves slow send mode